requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langchain-chroma==0.1.0
langchain-nvidia-ai-endpoints
chromadb
sentence-transformers
python-dotenv
pypdf
streamlit
